+++ packages/rrweb-snapshot/test/css.test.ts
@@ -1,4 +1,5 @@
 import { parse, Rule, Media } from '../src/css';
+import { fixSafariColons, escapeImportStatement } from './../src/utils';
 
 describe('css parser', () => {
   it('should save the filename and source', () => {
@@ -105,5 +106,76 @@ describe('css parser', () => {
     expect(rule.declarations!.length).toEqual(1);
     decl = rule.declarations![0];
     expect(decl.parent).toEqual(rule);
+  });
+
+  it('parses : in attribute selectors correctly', () => {
+    const out1 = fixSafariColons('[data-foo] { color: red; }');
+    expect(out1).toEqual('[data-foo] { color: red; }');
+
+    const out2 = fixSafariColons('[data-foo:other] { color: red; }');
+    expect(out2).toEqual('[data-foo\\:other] { color: red; }');
+
+    const out3 = fixSafariColons('[data-aa\\:other] { color: red; }');
+    expect(out3).toEqual('[data-aa\\:other] { color: red; }');
+  });
+
+  it('parses imports with quotes correctly', () => {
+    const out1 = escapeImportStatement({
+      cssText: `@import url("/foo.css;900;800"");`,
+      href: '/foo.css;900;800"',
+      media: {
+        length: 0,
+      },
+      layerName: null,
+      supportsText: null,
+    } as unknown as CSSImportRule);
+    expect(out1).toEqual(`@import url("/foo.css;900;800\\"");`);
+
+    const out2 = escapeImportStatement({
+      cssText: `@import url("/foo.css;900;800"") supports(display: flex);`,
+      href: '/foo.css;900;800"',
+      media: {
+        length: 0,
+      },
+      layerName: null,
+      supportsText: 'display: flex',
+    } as unknown as CSSImportRule);
+    expect(out2).toEqual(
+      `@import url("/foo.css;900;800\\"") supports(display: flex);`,
+    );
+
+    const out3 = escapeImportStatement({
+      cssText: `@import url("/foo.css;900;800"");`,
+      href: '/foo.css;900;800"',
+      media: {
+        length: 1,
+        mediaText: 'print, screen',
+      },
+      layerName: null,
+      supportsText: null,
+    } as unknown as CSSImportRule);
+    expect(out3).toEqual(`@import url("/foo.css;900;800\\"") print, screen;`);
+
+    const out4 = escapeImportStatement({
+      cssText: `@import url("/foo.css;900;800"") layer(layer-1);`,
+      href: '/foo.css;900;800"',
+      media: {
+        length: 0,
+      },
+      layerName: 'layer-1',
+      supportsText: null,
+    } as unknown as CSSImportRule);
+    expect(out4).toEqual(`@import url("/foo.css;900;800\\"") layer(layer-1);`);
+
+    const out5 = escapeImportStatement({
+      cssText: `@import url("/foo.css;900;800"") layer;`,
+      href: '/foo.css;900;800"',
+      media: {
+        length: 0,
+      },
+      layerName: '',
+      supportsText: null,
+    } as unknown as CSSImportRule);
+    expect(out5).toEqual(`@import url("/foo.css;900;800\\"") layer;`);
   });
 });