+++ packages/rrweb/src/record/mutation.ts
@@ -9,6 +9,7 @@ import {
   Mirror,
   isNativeShadowDom,
   getInputType,
+  toLowerCase,
 } from 'rrweb-snapshot';
 import type { observerParam, MutationBufferParam } from '../types';
 import type {
@@ -17,7 +18,6 @@ import type {
   attributeCursor,
   removedNodeMutation,
   addedNodeMutation,
-  styleAttributeValue,
   Optional,
 } from '@rrweb/types';
 import {
@@ -48,6 +48,7 @@ function isNodeInLinkedList(n: Node | NodeInLinkedList): n is NodeInLinkedList {
 class DoubleLinkedList {
   public length = 0;
   public head: DoubleLinkedListNode | null = null;
+  public tail: DoubleLinkedListNode | null = null;
 
   public get(position: number) {
     if (position >= this.length) {
@@ -95,6 +96,9 @@ class DoubleLinkedList {
       node.next = this.head;
       this.head = node;
     }
+    if (node.next === null) {
+      this.tail = node;
+    }
     this.length++;
   }
 
@@ -108,11 +112,15 @@ class DoubleLinkedList {
       this.head = current.next;
       if (this.head) {
         this.head.previous = null;
+      } else {
+        this.tail = null;
       }
     } else {
       current.previous.next = current.next;
       if (current.next) {
         current.next.previous = current.previous;
+      } else {
+        this.tail = current.previous;
       }
     }
     if (n.__ln) {
@@ -257,6 +265,7 @@ export default class MutationBuffer {
     // so that the mirror for takeFullSnapshot doesn't get mutated while it's event is being processed
 
     const adds: addedNodeMutation[] = [];
+    const addedIds = new Set<number>();
 
     /**
      * Sometimes child node may be pushed before its newly added
@@ -327,6 +336,7 @@ export default class MutationBuffer {
           nextId,
           node: sn,
         });
+        addedIds.add(sn.id);
       }
     };
 
@@ -368,8 +378,10 @@ export default class MutationBuffer {
         }
       }
       if (!node) {
-        for (let index = addList.length - 1; index >= 0; index--) {
-          const _node = addList.get(index);
+        let tailNode = addList.tail;
+        while (tailNode) {
+          const _node = tailNode;
+          tailNode = tailNode.previous;
           // ensure _node is defined before attempting to find value
           if (_node) {
             const parentId = this.mirror.getId(_node.value.parentNode);
@@ -424,13 +436,36 @@ export default class MutationBuffer {
           id: this.mirror.getId(text.node),
           value: text.value,
         }))
+        // no need to include them on added elements, as they have just been serialized with up to date attribubtes
+        .filter((text) => !addedIds.has(text.id))
         // text mutation's id was not in the mirror map means the target node has been removed
         .filter((text) => this.mirror.has(text.id)),
       attributes: this.attributes
-        .map((attribute) => ({
-          id: this.mirror.getId(attribute.node),
-          attributes: attribute.attributes,
-        }))
+        .map((attribute) => {
+          const { attributes } = attribute;
+          if (typeof attributes.style === 'string') {
+            const diffAsStr = JSON.stringify(attribute.styleDiff);
+            const unchangedAsStr = JSON.stringify(attribute._unchangedStyles);
+            // check if the style diff is actually shorter than the regular string based mutation
+            // (which was the whole point of #464 'compact style mutation').
+            if (diffAsStr.length < attributes.style.length) {
+              // also: CSSOM fails badly when var() is present on shorthand properties, so only proceed with
+              // the compact style mutation if these have all been accounted for
+              if (
+                (diffAsStr + unchangedAsStr).split('var(').length ===
+                attributes.style.split('var(').length
+              ) {
+                attributes.style = attribute.styleDiff;
+              }
+            }
+          }
+          return {
+            id: this.mirror.getId(attribute.node),
+            attributes: attributes,
+          };
+        })
+        // no need to include them on added elements, as they have just been serialized with up to date attribubtes
+        .filter((attribute) => !addedIds.has(attribute.id))
         // attribute mutation's id was not in the mirror map means the target node has been removed
         .filter((attribute) => this.mirror.has(attribute.id)),
       removes: this.removes,
@@ -462,7 +497,14 @@ export default class MutationBuffer {
     if (isIgnored(m.target, this.mirror)) {
       return;
     }
-    const unattachedDoc = new Document(); // avoid upsetting original document from a Content Security point of view
+    let unattachedDoc;
+    try {
+      // avoid upsetting original document from a Content Security point of view
+      unattachedDoc = document.implementation.createHTMLDocument();
+    } catch (e) {
+      // fallback to more direct method
+      unattachedDoc = this.doc;
+    }
     switch (m.type) {
       case 'characterData': {
         const value = m.target.textContent;
@@ -530,6 +572,8 @@ export default class MutationBuffer {
           item = {
             node: m.target,
             attributes: {},
+            styleDiff: {},
+            _unchangedStyles: {},
           };
           this.attributes.push(item);
         }
@@ -544,46 +588,43 @@ export default class MutationBuffer {
           target.setAttribute('data-rr-is-password', 'true');
         }
 
-        if (attributeName === 'style') {
-          const old = unattachedDoc.createElement('span');
-          if (m.oldValue) {
-            old.setAttribute('style', m.oldValue);
-          }
-          if (
-            item.attributes.style === undefined ||
-            item.attributes.style === null
-          ) {
-            item.attributes.style = {};
-          }
-          const styleObj = item.attributes.style as styleAttributeValue;
-          for (const pname of Array.from(target?.style || [])) {
-            const newValue = target.style?.getPropertyValue(pname);
-            const newPriority = target.style?.getPropertyPriority(pname);
-            if (
-              newValue !== old.style?.getPropertyValue(pname) ||
-              newPriority !== old.style?.getPropertyPriority(pname)
-            ) {
-              if (newPriority === '') {
-                styleObj[pname] = newValue;
-              } else {
-                styleObj[pname] = [newValue, newPriority];
-              }
-            }
-          }
-          for (const pname of Array.from(old?.style || [])) {
-            if (target.style?.getPropertyValue(pname) === '') {
-              // "if not set, returns the empty string"
-              styleObj[pname] = false; // delete
-            }
-          }
-        } else if (!ignoreAttribute(target.tagName, attributeName, value)) {
+        if (!ignoreAttribute(target.tagName, attributeName, value)) {
           // overwrite attribute if the mutations was triggered in same time
           item.attributes[attributeName] = transformAttribute(
             this.doc,
-            target.tagName,
-            attributeName,
+            toLowerCase(target.tagName),
+            toLowerCase(attributeName),
             value,
           );
+          if (attributeName === 'style') {
+            const old = unattachedDoc.createElement('span');
+            if (m.oldValue) {
+              old.setAttribute('style', m.oldValue);
+            }
+            for (const pname of Array.from(target?.style || [])) {
+              const newValue = target.style?.getPropertyValue(pname);
+              const newPriority = target.style?.getPropertyPriority(pname);
+              if (
+                newValue !== old.style?.getPropertyValue(pname) ||
+                newPriority !== old.style?.getPropertyPriority(pname)
+              ) {
+                if (newPriority === '') {
+                  item.styleDiff[pname] = newValue;
+                } else {
+                  item.styleDiff[pname] = [newValue, newPriority];
+                }
+              } else {
+                // for checking
+                item._unchangedStyles[pname] = [newValue, newPriority];
+              }
+            }
+            for (const pname of Array.from(old?.style || [])) {
+              if (target.style?.getPropertyValue(pname) === '') {
+                // "if not set, returns the empty string"
+                item.styleDiff[pname] = false; // delete
+              }
+            }
+          }
         }
         break;
       }