+++ packages/rrweb/src/utils.ts
@@ -171,6 +171,15 @@ export function patch(
   }
 }
 
+// guard against old third party libraries which redefine Date.now
+let nowTimestamp = Date.now;
+
+if (!(/*@__PURE__*/ /[1-9][0-9]{12}/.test(Date.now().toString()))) {
+  // they have already redefined it! use a fallback
+  nowTimestamp = () => new Date().getTime();
+}
+export { nowTimestamp };
+
 export function getWindowScroll(win: (Window & typeof globalThis) | null): {
   left: number;
   top: number;